fix: 単価マスタで小数点の単価を保存できない不具合を修正
unit_prices.price_yenがinteger型だったため、205.7円のような小数を
含む単価の保存時に "invalid input syntax for type integer" エラーと
なっていた。numeric(10,2)に変更し、小数点を許可する。

+++ supabase/migrations/20260729093619_create_master_data.sql
@@ -285,7 +285,7 @@ create table public.unit_prices (
   area_id uuid not null references public.areas (id),
   delivery_type_id uuid not null references public.delivery_types (id),
   effective_from date not null,
-  price_yen integer not null,
+  price_yen numeric(10, 2) not null,
   created_by uuid references public.profiles (id),
   created_at timestamptz not null default now(),
   unique (price_kind, area_id, delivery_type_id, effective_from)
